Addressed pr changes!
