+++ bgls/__init__.py
@@ -21,6 +21,5 @@ from bgls.utils import (
     cirq_density_matrix_bitstring_probability,
     cirq_stabilizer_ch_bitstring_probability,
     act_on_near_clifford,
-    cirq_mps_bitstring_probability,
     generate_random_circuit,
 )
