+++ bgls/utils.py
@@ -228,25 +228,3 @@ def generate_random_circuit(
         moments.append(cirq.circuits.Moment(operations))
 
     return cirq.circuits.Circuit(moments)
-
-
-def cirq_mps_bitstring_probability(
-    mps: cirq.contrib.quimb.MPSState, bitstring: str
-) -> float:
-    """
-    Returns the probability of measuring the `bitstring` (|z⟩) in the
-    'cirq.contrib.quimb.MPSState' mps.
-    Args:
-        mps: Matrix Product State as a 'cirq.contrib.quimb.MPSState'.
-        bitstring: Bitstring |z⟩ as a binary string.
-    """
-    M_subset = []
-    for i, Ai in enumerate(mps.M):
-        qubit_index = mps.i_str(i)
-        # selecting the component with matching bitstring:
-        A_subset = Ai.isel({qubit_index: int(bitstring[i])})
-        M_subset.append(A_subset)
-
-    tensor_network = qtn.TensorNetwork(M_subset)
-    state_vector = tensor_network.contract(inplace=False)
-    return np.power(np.abs(state_vector), 2)
