+++ requirements.txt
@@ -1,4 +1,1 @@
 cirq-core>=1.1.0
-quimb
-opt_einsum
-autoray
